Move implementation of macro to separate file
